# Add batch_converter hook + multiprocessing support to AnnLoader