--- a/.vscode/settings.json
+++ b/.vscode/settings.json
@@ -24,4 +24,5 @@
         //"-nauto",
     ],
     "python.terminal.activateEnvironment": true,
+    "cursorpyright.analysis.typeCheckingMode": "basic",
 }