--- a/src/anndata/experimental/pytorch/__init__.py
+++ b/src/anndata/experimental/pytorch/__init__.py
@@ -1,5 +1,10 @@
 from __future__ import annotations
 
+# public re-exports
 from ._annloader import AnnLoader
+from .converters import to_tensor_dict as batch_dict_converter
 
-__all__ = ["AnnLoader"]
+__all__: list[str] = [
+    "AnnLoader",
+    "batch_dict_converter",
+]